+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- API responses now exposed under specs, rather than internal packages.
+- Added `RESPONSE_CLIENT_CONNECT` message sent to a host when a new client connects.
 
 ## [0.2.0] - 2020-03-24
 ### Added