+++ CHANGELOG.md
@@ -6,11 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.2.0] - 2020-03-24
+## [0.3.0] - 2021-14-16
+### Changed
+- API responses now exposed under specs, rather than internal packages.
+- Added `RESPONSE_CLIENT_CONNECT` message sent to a host when a new client connects.
+
+## [0.2.0] - 2021-03-24
 ### Added
 - Publishing protobuf specification files.
 
-## [0.1.0] - 2020-03-23
+## [0.1.0] - 2021-03-23
 ### Added
 - Initial release of jamjar-relay-server